clean pargraphics

+++ OpenGl1_Editor/src/ChokoLait.cpp
@@ -139,7 +139,7 @@ void ChokoLait::MouseGL(GLFWwindow* window, int button, int state, int mods) {
 }
 
 void ChokoLait::MouseScrGL(GLFWwindow* window, double xoff, double yoff) {
-	Input::_mouseScroll = yoff;
+	Input::_mouseScroll = (float)yoff;
 }
 
 void ChokoLait::MotionGL(GLFWwindow* window, double x, double y) {
