+++ OpenGl1_Editor/src/md/ParMenu.cpp
@@ -28,19 +28,19 @@ void ParMenu::InitList() {
 }
 
 void ParMenu::Draw() {
-	Engine::DrawQuad(0, 0, expandPos, Display::height, white(0.9f, 0.15f));
+	Engine::DrawQuad(0, 0, expandPos, (float)Display::height, white(0.9f, 0.15f));
 	if (expanded) {
 		float f = 20;
-		Engine::BeginStencil(0, 0, expandPos, Display::height);
+		Engine::BeginStencil(0, 0, expandPos, (float)Display::height);
 
 
 		Engine::EndStencil();
 
 		for (uint i = 0; i < 4; i++) {
 			if (i == activeMenu)
-				Engine::DrawQuad(expandPos, 81 * i, 17, 81, white(0.9f, 0.15f));
+				Engine::DrawQuad(expandPos, 81.0f * i, 17, 81, white(0.9f, 0.15f));
 			else
-				if (Engine::Button(expandPos, 81 * i, 16, 80, white(0.7f, 0.1f), white(1, 0.2f), white(1, 0.05f)) == MOUSE_RELEASE) {
+				if (Engine::Button(expandPos, 81.0f * i, 16, 80, white(0.7f, 0.1f), white(1, 0.2f), white(1, 0.05f)) == MOUSE_RELEASE) {
 					activeMenu = i;
 				}
 		}
@@ -59,17 +59,17 @@ void ParMenu::Draw() {
 		font->Align(ALIGN_TOPLEFT);
 		Engine::ResetUIMatrix();
 
-		Engine::DrawQuad(expandPos, Display::height - 16, 16, 16, white(0.9f, 0.15f));
-		if (Engine::Button(expandPos, Display::height - 16, 16, 16, Icons::collapse, white(0.8f), white(), white(0.5f)) == MOUSE_RELEASE)
+		Engine::DrawQuad(expandPos, Display::height - 16.0f, 16, 16, white(0.9f, 0.15f));
+		if (Engine::Button(expandPos, Display::height - 16.0f, 16, 16, Icons::collapse, white(0.8f), white(), white(0.5f)) == MOUSE_RELEASE)
 			expanded = false;
 		Engine::EndStencil();
 		expandPos = min(expandPos + 1500 * Time::delta, 150.0f);
 	}
 	else {
-		if (Engine::Button(expandPos, Display::height - 16, 115, 16, white(0.9f, 0.15f), white(1, 0.15f), white(1, 0.05f)) == MOUSE_RELEASE)
+		if (Engine::Button(expandPos, Display::height - 16.0f, 115, 16, white(0.9f, 0.15f), white(1, 0.15f), white(1, 0.05f)) == MOUSE_RELEASE)
 			expanded = true;
-		UI::Texture(expandPos, Display::height - 16, 16, 16, Icons::expand);
-		UI::Label(expandPos + 18, Display::height - 15, 12, "Particle View", font, white());
+		UI::Texture(expandPos, Display::height - 16.0f, 16, 16, Icons::expand);
+		UI::Label(expandPos + 18, Display::height - 15.0f, 12, "Particle View", font, white());
 		expandPos = max(expandPos - 1500 * Time::delta, 0.0f);
 	}
 }
